ml-training/requirements.txt
# ML Training Dependencies for SatsVerdant

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
imgaug==0.4.0

# ML Utilities
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Experiment Tracking
mlflow==2.9.2

# Data Augmentation & Synthetic Generation
albumentations==1.3.1
rembg==2.0.50

# Dataset Tools
taco==0.1.0
kaggle==1.5.16

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
